Change argument type to double
DBus does not support float. This was not causing any error, because it
got implicitly converted.

+++ src/plugins/simulator/xplane/xswiftbustrafficproxy.cpp
@@ -62,7 +62,7 @@ namespace BlackSimPlugin
             m_dbusInterface->callDBus(QLatin1String("setMaxPlanes"), planes);
         }
 
-        void CXSwiftBusTrafficProxy::setMaxDrawDistance(float nauticalMiles)
+        void CXSwiftBusTrafficProxy::setMaxDrawDistance(double nauticalMiles)
         {
             m_dbusInterface->callDBus(QLatin1String("setMaxDrawDistance"), nauticalMiles);
         }
